+++ capstone_project/Tests/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="FluentAssertions" version="4.2.2" targetFramework="net452" />
-  <package id="Selenium.WebDriver" version="2.53.0" targetFramework="net452" />
 </packages>
