+++ capstone_project/Web/Views/Shared/_Header.cshtml
@@ -1,5 +1,5 @@
 ﻿<div class="navbar navbar-default navbar-fixed-top">
-    <div class="container" name="nav-links">
+    <div class="container">
         <div class="row">
         <div class="navbar-header">
             <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
